Crate sc_network_transactions

Transactions handling to plug on top of the network service.

Usage:

Use TransactionsHandlerPrototype::new to create a prototype.

Pass the NonDefaultSetConfig returned from TransactionsHandlerPrototype::new to the network configuration as an extra peers set.

Use TransactionsHandlerPrototype::build then TransactionsHandler::run to obtain a Future that processes transactions.

Modules

config

Configuration of the transaction protocol

Structs

TransactionsHandler

Handler for transactions. Call TransactionsHandler::run to start the processing.

TransactionsHandlerController

Controls the behaviour of a TransactionsHandler it is connected to.

TransactionsHandlerPrototype

Prototype for a TransactionsHandler.

Type Aliases

Transactions

A set of transactions.