Crate sc_network_transactions

Transactions handling to plug on top of the network service.

Usage:

Use TransactionsHandlerPrototype::new to create a prototype.

Pass the return value of TransactionsHandlerPrototype::set_config to the network configuration as an extra peers set.

Use TransactionsHandlerPrototype::build then TransactionsHandler::run to obtain a Future that processes transactions.

Modules

config

Configuration of the transaction protocol

Structs

TransactionsHandler

Handler for transactions. Call TransactionsHandler::run to start the processing.

TransactionsHandlerController

Controls the behaviour of a TransactionsHandler it is connected to.

TransactionsHandlerPrototype

Prototype for a TransactionsHandler.

Type Definitions

Transactions

A set of transactions.